refactor: Removed parliamentary codes from TableCompositionPage
