+++ src/page/TableComposition/components/EditDeleteButtons/EditDeleteButtons.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 import EditDeleteButtons from './EditDeleteButtons';
-import { ParliamentaryTimeBoxInfo } from '../../../../type/type';
+import { TimeBoxInfo } from '../../../../type/type';
 
 const meta: Meta<typeof EditDeleteButtons> = {
   title: 'page/TableSetup/components/EditDeleteButtons',
@@ -12,16 +12,37 @@ export default meta;
 
 type Story = StoryObj<typeof EditDeleteButtons>;
 
-const mockInfo: ParliamentaryTimeBoxInfo = {
+const normalTimeBoxInfo: TimeBoxInfo = {
   stance: 'PROS',
-  type: 'OPENING',
-  time: 150,
-  speakerNumber: 1,
+  speechType: '교차 조사',
+  boxType: 'NORMAL',
+  time: 120,
+  speaker: '김철수 토론자',
+  timePerSpeaking: null,
+  timePerTeam: null,
 };
 
-export const Default: Story = {
+export const NormalTimeBox: Story = {
   args: {
-    info: mockInfo,
+    info: normalTimeBoxInfo,
+    onSubmitEdit: () => {},
+    onSubmitDelete: () => {},
+  },
+};
+
+const timeBasedTimeBoxInfo: TimeBoxInfo = {
+  stance: 'NEUTRAL',
+  speechType: '자유토론',
+  boxType: 'TIME_BASED',
+  time: null,
+  speaker: null,
+  timePerSpeaking: 120,
+  timePerTeam: 480,
+};
+
+export const TimeBasedTimeBox: Story = {
+  args: {
+    info: timeBasedTimeBoxInfo,
     onSubmitEdit: () => {},
     onSubmitDelete: () => {},
   },
