+++ src/page/TableComposition/components/TimeBoxStep/TimeBoxStep.tsx
@@ -1,29 +1,19 @@
 import DebatePanel from '../DebatePanel/DebatePanel';
-import CustomizeDebatePanel from '../DebatePanel/CustomizeDebatePanel';
 import TimerCreationButton from '../TimerCreationButton/TimerCreationButton';
-import TimerCreationContent from '../TimerCreationContent/TimerCreationContent';
 import { useModal } from '../../../../hooks/useModal';
-import { ParliamentaryTimeBoxInfo } from '../../../../type/type';
 import { useDragAndDrop } from '../../../../hooks/useDragAndDrop';
 import DefaultLayout from '../../../../layout/defaultLayout/DefaultLayout';
 import PropsAndConsTitle from '../../../../components/ProsAndConsTitle/PropsAndConsTitle';
-import { TableFormData, TimeBoxInfo } from '../../hook/useTableFrom';
 import HeaderTableInfo from '../../../../components/HeaderTableInfo/HeaderTableInfo';
 import HeaderTitle from '../../../../components/HeaderTitle/HeaderTitle';
-import { CustomizeTimeBoxInfo } from '../../../../type/type';
-import CustomizeTimerCreationContent from '../TimerCreationContent/CustomizeTimerCreationContent';
+import TimerCreationContent from '../TimerCreationContent/TimerCreationContent';
+import { TableFormData, TimeBoxInfo } from '../../../../type/type';
 
 interface TimeBoxStepProps {
   initData: TableFormData;
   onTimeBoxChange: React.Dispatch<React.SetStateAction<TimeBoxInfo[]>>;
   onButtonClick: () => void;
   isEdit?: boolean;
-}
-// customize 타입가드
-function isCustomize(
-  info: TableFormData['info'],
-): info is Extract<TableFormData, { info: { type: 'CUSTOMIZE' } }>['info'] {
-  return info.type === 'CUSTOMIZE';
 }
 
 export default function TimeBoxStep(props: TimeBoxStepProps) {
@@ -54,26 +44,14 @@ export default function TimeBoxStep(props: TimeBoxStepProps) {
 
   const isAbledSummitButton = initTimeBox.length !== 0;
 
-  // 타임박스 렌더링 분기 처리
   const renderTimeBoxItem = (info: TimeBoxInfo, index: number) => {
-    if (isCustomize(initData.info)) {
-      return (
-        <CustomizeDebatePanel
-          key={index}
-          info={info as CustomizeTimeBoxInfo}
-          onSubmitEdit={(updatedInfo) => handleSubmitEdit(index, updatedInfo)}
-          prosTeamName={initData.info.prosTeamName}
-          consTeamName={initData.info.consTeamName}
-          onSubmitDelete={() => handleSubmitDelete(index)}
-          onMouseDown={() => handleMouseDown(index)}
-        />
-      );
-    }
     return (
       <DebatePanel
         key={index}
-        info={info as ParliamentaryTimeBoxInfo}
+        info={info as TimeBoxInfo}
         onSubmitEdit={(updatedInfo) => handleSubmitEdit(index, updatedInfo)}
+        prosTeamName={initData.info.prosTeamName}
+        consTeamName={initData.info.consTeamName}
         onSubmitDelete={() => handleSubmitDelete(index)}
         onMouseDown={() => handleMouseDown(index)}
       />
@@ -84,10 +62,7 @@ export default function TimeBoxStep(props: TimeBoxStepProps) {
     <DefaultLayout>
       <DefaultLayout.Header>
         <DefaultLayout.Header.Left>
-          <HeaderTableInfo
-            name={initData.info.name}
-            type={initData.info.type}
-          />
+          <HeaderTableInfo name={initData.info.name} />
         </DefaultLayout.Header.Left>
         <DefaultLayout.Header.Center>
           <HeaderTitle title={initData.info.agenda} />
@@ -97,14 +72,10 @@ export default function TimeBoxStep(props: TimeBoxStepProps) {
 
       <DefaultLayout.ContentContainer>
         <section className="mx-auto flex w-full max-w-4xl flex-col justify-center">
-          {isCustomize(initData.info) ? (
-            <PropsAndConsTitle
-              prosTeamName={initData.info.prosTeamName}
-              consTeamName={initData.info.consTeamName}
-            />
-          ) : (
-            <PropsAndConsTitle />
-          )}
+          <PropsAndConsTitle
+            prosTeamName={initData.info.prosTeamName}
+            consTeamName={initData.info.consTeamName}
+          />
 
           <DragAndDropWrapper>
             {initTimeBox.map((info, index) => (
@@ -132,33 +103,17 @@ export default function TimeBoxStep(props: TimeBoxStepProps) {
         </div>
       </DefaultLayout.StickyFooterWrapper>
 
-      {isCustomize(initData.info) ? (
-        <ModalWrapper closeButtonColor="text-neutral-1000">
-          <CustomizeTimerCreationContent
-            beforeData={
-              initTimeBox[initTimeBox.length - 1] as CustomizeTimeBoxInfo
-            }
-            prosTeamName={initData.info.prosTeamName}
-            consTeamName={initData.info.consTeamName}
-            onSubmit={(data) => {
-              onTimeBoxChange((prev) => [...prev, data]);
-            }}
-            onClose={closeModal}
-          />
-        </ModalWrapper>
-      ) : (
-        <ModalWrapper>
-          <TimerCreationContent
-            beforeData={
-              initTimeBox[initTimeBox.length - 1] as ParliamentaryTimeBoxInfo
-            }
-            onSubmit={(data) => {
-              onTimeBoxChange((prev) => [...prev, data]);
-            }}
-            onClose={closeModal}
-          />
-        </ModalWrapper>
-      )}
+      <ModalWrapper closeButtonColor="text-neutral-1000">
+        <TimerCreationContent
+          beforeData={initTimeBox[initTimeBox.length - 1] as TimeBoxInfo}
+          prosTeamName={initData.info.prosTeamName}
+          consTeamName={initData.info.consTeamName}
+          onSubmit={(data) => {
+            onTimeBoxChange((prev) => [...prev, data]);
+          }}
+          onClose={closeModal}
+        />
+      </ModalWrapper>
     </DefaultLayout>
   );
 }
