+++ src/page/TableComposition/hook/useTableFrom.tsx
@@ -2,28 +2,9 @@ import { useEffect } from 'react';
 import { useNavigate, useNavigationType } from 'react-router-dom';
 import { TableCompositionStep } from '../TableComposition';
 import useBrowserStorage from '../../../hooks/useBrowserStorage';
-import {
-  ParliamentaryTimeBoxInfo,
-  CustomizeTimeBoxInfo,
-  ParliamentaryInfo,
-  CustomizeInfo,
-} from '../../../type/type';
-import useAddParliamentaryTable from '../../../hooks/mutations/useAddParliamentaryDebateTable';
-import { usePutParliamentaryDebateTable } from '../../../hooks/mutations/usePutParliamentaryDebateTable';
+import { DebateInfo, TableFormData, TimeBoxInfo } from '../../../type/type';
 import useAddCustomizeTable from '../../../hooks/mutations/useAddCustomizeDebateTable';
 import { usePutCustomizeDebateTable } from '../../../hooks/mutations/usePutCustomizeDebateTable';
-
-// type 필드 포함한 인터페이스를 info로 사용. 의회식/사용자지정 구분
-interface ParliamentaryTableFormData {
-  info: ParliamentaryInfo;
-  table: ParliamentaryTimeBoxInfo[];
-}
-interface CustomizeTableFormData {
-  info: CustomizeInfo;
-  table: CustomizeTimeBoxInfo[];
-}
-export type TableFormData = ParliamentaryTableFormData | CustomizeTableFormData;
-export type TimeBoxInfo = ParliamentaryTimeBoxInfo | CustomizeTimeBoxInfo;
 
 const useTableFrom = (
   currentStep: TableCompositionStep,
@@ -40,7 +21,6 @@ const useTableFrom = (
         info: {
           name: '',
           agenda: '',
-          type: 'CUSTOMIZE',
           prosTeamName: '',
           consTeamName: '',
           warningBell: true,
@@ -63,8 +43,8 @@ const useTableFrom = (
     if (initData) {
       setFormData(initData);
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [initData]);
+    // Originaly here was exhaustive-deps
+  }, [initData, setFormData]);
 
   useEffect(() => {
     if (currentStep === 'TimeBox' && navigationType === 'POP') {
@@ -77,34 +57,19 @@ const useTableFrom = (
   > = (action) => {
     setFormData((prev) => {
       const newInfo = typeof action === 'function' ? action(prev.info) : action;
+      const debateInfo: DebateInfo = {
+        name: newInfo.name,
+        agenda: newInfo.agenda,
+        warningBell: newInfo.warningBell,
+        finishBell: newInfo.finishBell,
+        prosTeamName: newInfo.prosTeamName,
+        consTeamName: newInfo.consTeamName,
+      };
 
-      if (isCustomizeInfo(newInfo)) {
-        const customizeInfo: CustomizeInfo = {
-          name: newInfo.name,
-          agenda: newInfo.agenda,
-          type: 'CUSTOMIZE',
-          warningBell: newInfo.warningBell,
-          finishBell: newInfo.finishBell,
-          prosTeamName: newInfo.prosTeamName,
-          consTeamName: newInfo.consTeamName,
-        };
-        return {
-          info: customizeInfo,
-          table: prev.table as CustomizeTimeBoxInfo[],
-        };
-      } else {
-        const parliamentaryInfo: ParliamentaryInfo = {
-          name: newInfo.name,
-          agenda: newInfo.agenda,
-          type: 'PARLIAMENTARY',
-          warningBell: newInfo.warningBell,
-          finishBell: newInfo.finishBell,
-        };
-        return {
-          info: parliamentaryInfo,
-          table: prev.table as ParliamentaryTimeBoxInfo[],
-        };
-      }
+      return {
+        info: debateInfo,
+        table: prev.table as TimeBoxInfo[],
+      };
     });
   };
 
@@ -114,76 +79,36 @@ const useTableFrom = (
     setFormData((prev) => {
       const newTable =
         typeof action === 'function' ? action(prev.table) : action;
-
-      if (isCustomizeInfo(prev.info)) {
-        return {
-          info: prev.info,
-          table: newTable,
-        } as CustomizeTableFormData;
-      } else {
-        return {
-          info: prev.info,
-          table: newTable,
-        } as ParliamentaryTableFormData;
-      }
+      return {
+        info: prev.info,
+        table: newTable,
+      } as TableFormData;
     });
   };
 
-  const { mutate: addParliamentary } = useAddParliamentaryTable((tableId) => {
-    removeValue();
-    navigate(`/overview/parliamentary/${tableId}`);
-  });
-
-  const { mutate: addCustomize } = useAddCustomizeTable((tableId) => {
+  const { mutate: onAddTable } = useAddCustomizeTable((tableId) => {
     removeValue();
     navigate(`/overview/customize/${tableId}`);
   });
 
-  const { mutate: editParliamentary } = usePutParliamentaryDebateTable(
-    (tableId) => {
-      removeValue();
-      navigate(`/overview/parliamentary/${tableId}`);
-    },
-  );
-
-  const { mutate: editCustomize } = usePutCustomizeDebateTable((tableId) => {
+  const { mutate: onModifyTable } = usePutCustomizeDebateTable((tableId) => {
     removeValue();
     navigate(`/overview/customize/${tableId}`);
   });
-
-  // customize 타입 가드
-  function isCustomizeInfo(info: TableFormData['info']): info is CustomizeInfo {
-    return info.type === 'CUSTOMIZE';
-  }
 
   const AddTable = () => {
-    if (isCustomizeInfo(formData.info)) {
-      addCustomize({
-        info: formData.info,
-        table: formData.table as CustomizeTimeBoxInfo[],
-      });
-    } else {
-      addParliamentary({
-        info: formData.info,
-        table: formData.table as ParliamentaryTimeBoxInfo[],
-      });
-    }
+    onAddTable({
+      info: formData.info,
+      table: formData.table as TimeBoxInfo[],
+    });
   };
 
   const EditTable = (tableId: number) => {
-    if (isCustomizeInfo(formData.info)) {
-      editCustomize({
-        tableId,
-        info: formData.info,
-        table: formData.table as CustomizeTimeBoxInfo[],
-      });
-    } else {
-      editParliamentary({
-        tableId,
-        info: formData.info,
-        table: formData.table as ParliamentaryTimeBoxInfo[],
-      });
-    }
+    onModifyTable({
+      tableId,
+      info: formData.info,
+      table: formData.table as TimeBoxInfo[],
+    });
   };
 
   return {
