+++ src/page/TableOverviewPage/TableOverview.tsx
@@ -6,7 +6,7 @@ import HeaderTitle from '../../components/HeaderTitle/HeaderTitle';
 import { RiEditFill, RiSpeakFill } from 'react-icons/ri';
 import usePatchCustomizeTable from '../../hooks/mutations/usePatchCustomizeDebateTable';
 import { useGetCustomizeTableData } from '../../hooks/query/useGetCustomizeTableData';
-import CustomizeDebatePanel from '../TableComposition/components/DebatePanel/CustomizeDebatePanel';
+import DebatePanel from '../TableComposition/components/DebatePanel/DebatePanel';
 import { useTableShare } from '../../hooks/useTableShare';
 import { MdOutlineIosShare } from 'react-icons/md';
 import { StanceToString } from '../../type/type';
@@ -56,7 +56,7 @@ export default function TableOverview() {
 
             <div className="flex w-full flex-col gap-2">
               {data?.table.map((info, index) => (
-                <CustomizeDebatePanel
+                <DebatePanel
                   key={index}
                   info={info}
                   prosTeamName={data.info.prosTeamName}
